backend/requirements.txt
fastapi
uvicorn
python-socketio
openai
ollama
python-dotenv
requests
hypercorn
